# backend/swagger.yaml
...
openapi: 3.0.0
info:
  title: Equipment Management API
  version: 2.0.0
  description: API для системы учета оборудования
servers:
  - url: http://localhost:3000/api
    description: Development server
paths:
  /nodes:
    get:
      summary: Получить список узлов
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: sort
          in: query
          description: 'Сортировка: -created_at (по убыванию) или name,-status'
          schema: { type: string, default: '-created_at' }
        - name: filters
          in: query
          description: 'JSON фильтров: {"status": "исправен", "manufacturer": {"operator": "like", "value": "НПП"}}'
          schema: { type: string }
        - name: columns
          in: query
          description: 'Выбор колонок через запятую: node_id,manufacturer,model'
          schema: { type: string }
      responses:
        200:
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { type: array }
                  meta:
                    type: object
                    properties:
                      total: { type: integer }
                      page: { type: integer }
                      limit: { type: integer }
                      pages: { type: integer }
    post:
      summary: Создать узел
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [node_type_id, manufacturer, model, status, location, subsystem_id]
              properties:
                node_type_id: { type: string, format: uuid }
                manufacturer: { type: string }
                model: { type: string }
                status: { type: string, enum: ['получен (не проверен)', 'исправен', 'неисправен', 'в ремонте', 'на поверке', 'законсервирован', 'списан'] }
                location: { type: string }
                subsystem_id: { type: string, format: uuid }
                parameters: { type: object }
      responses:
        201:
          description: Создано успешно
...